11 radioisotope thermoelectric generators
This looked like a graph problem almost immediately but it took ages to
just write and debug tiny things.

Move microchips and generators in the elevator, with the given
constraints. Each state (= where each object is) is a node, and
movements using the elevator are edges. Then just breadth-first-search
it into the end state, saving distances on the go.

The state is a neat bit string, with four bits for each object
representing the current floor.

Messy code because debugging and visualization stuff.