14 only first triplet, should've read the spec
So only consider the first triplet, don't store all of them. Read the
spec wrong somehow.

How the prev commit works: Use plain byte arrays to scan for
"characters" in the "hex" output, although it wouldn't probably matter
that much for speed. Store the hashes in a vector because recomputing
may not be fast (and oldest ones could be deleted, but couldn't care)
although it probably is really fast. Premature stupidity, bleh.