25 clock signal
Lots of copy pasta again. A new instruction is added to output the value
of a register. The outputs of the program seem to always be zeros and
ones, just not in the right format if the reg a isn't correct.

I dumped some debug messages and it seems that the code repeats blocks
of length 12, regardless of the first reg state. Anything longer would
work, too; this is just the minimal (fastest) one.

An automatic way for finding the block size would be to track the states
in executing the thing and checking when we get to one that has been
already visited.

There is no part two for today.

Might look back at the assembunny sources of each puzzle now soon and
"disassemble" them for fun.