fix 19b perf: use linked lists, keep pointers
The massive memcpys with the list were a silly thing to do, and there's
a pretty neat way to solve this in any case nearly instantly. When a
victim node is deleted from an even-sized list, the victim moves to the
one next to that. For odd-sized, the victim moves by two. Pen and paper
ftw. I knew it.

Track the node just before the victim, and current stealer to easily
find out which won.