16b dance as a whole
Do the dance moves for 1000000000 times. Executing all those moves again
and again would take ages, but there's a catch: after a while, the
programs restore to their original order, and the length of this cycle
depends on the largest common multiple of all internal cycles of single
positions.

Embarrassingly, I tried something else first and failed miserably.
Intuitively, the programs always change in the same way for each
execution of the dance moves, and collapsing the whole dance to only
executing that permutation would be much faster. However, this would
only work without the "partner" move, which is not index-based.