23b time to fix the problem
Yay, assembly code reverse engineering! The naive program overheats
because the algorithm is badly designed, but just looping a little less
makes it faster. I translated it to plain Rust as well.

This actually computes the count of non-primes between 107900 and 124900
but stepping 17 at a time (for my input), so there would be immensely
faster algos for that, but this runs in like three seconds which is
enough.