cleanup 24 for both parts
Parts 1 and 2 use a similar search in the domino component graph; one
needs just the strongest path, and the other the longest (taking the
strongest if there are ties). Drop the copypasta and generalize the
search.

The search takes into account the two-way nature of the nodes (done in
earlier commits already): when going in from one side, only the
connections to the other side are considered.

Another neat way to approach this problem could be to model each
component as multiple nodes in the graph somehow, so that when going in
to one node the only next direction would be out of the other side. Then
the basic dfs would be easier, requiring more heavy connectivity setup.