cleanup 20 and fix corner cases
I got the correct result even with bugs, and by some magical chance the
escaper-deletion dropping accidentally the last element in the list
instead of the correct particle still worked??! Also other bugs, oh so
many. Now the escaper-comparison needs to be careful when searching the
max element etc.

Instead of checking escaping particles, the sim could just run "long
enough" which seems to be not even that many iterations, but how could
you be sure then?