6 chronal coordinates
Just brute-force the map around the given coordinates and find the
maximum area as specified. The infinite areas aren't filtered out
because the largest one even without filtering happened to be the
correct one, although that's not necessarily true for all inputs. I'm
too lazy for this and running late, it's day 13 already.