8b value of the root node
Now the value of a node that has children is the sum of such children
that are indexed by metadata, and there can be duplicates. Node with no
children has a value of just the metadata sum; the recursion ends there.

Not sure how effective this iterator dance is, but with raw vectors we'd
still need to skip children like that. Trivially fast even without a
clever lookup structure.