10, 10b: the stars align
I accidentally both parts in a single go. Clearly the stars are coming
together; instead of doing some clever math to calculate the exact time,
iterate the physics a second at a time and see when the bounding box
starts to grow again. The text is assumed to take little space compared
to when the stars are inflated around the space; the smallest space the
stars can fit in is the text.

Draw a picture when we've found the correct timestep; I'm not doing any
OCR for this. By chance the filename contains the answer to the second
part.