better 11b, compute incrementally
Instead of redoing sums of smaller squares, make a compromise and still
calculate all sizes on each corner separately but grow the size
incrementally and only add the newly appeared bottom and right sides of
the square.

Some kind of BSP structure might be even better, but this runs in 13
seconds including all the test cases too. Good enough.