12 subterranean sustainability
After exploring a little, we find yet another pattern substitution thing
like I remember from last year perhaps. I'd love to do this with some
sort of bit string library, but let's see if arrays of u8s are
performant enough.

Read the patterns that produce a plant in the pot and discard the
others. When generating the offspring pots, check if the current pattern
of flowers is found in the rulebook.