12b another 1500 years and more
Fifty billion generations is too much to compute like that. After a log
of 200 or so generations it seems clear that there will be a static
pattern that simply moves to the right. Moving the same amount of pots
every time contributes a constant difference to the score, so we can
look ahead and multiply the iteration count by that to get the final
score.