13 mine cart madness
This crop requires significant logistics and the logicstics requires
significant extra code now as well but after the basic stuff is in place
this simulation is rather nice. The moving directions and carts have
proper data types and the map is still those characters, although this
would be curiously clever if those were plain numbers.

The debug dump ansi-colors the carts so they stand out. An animation
would be fun.