14b got the plan backwards
Loop until some particular set of digits is found. Perhaps comparing a
slice to a list of digits instead of computing a number would be easier
than this - now the size of the list of digits is also necessary.

The values are in u8 this time because previously this had a bug and I
thought the puzzle would just need ridiculous amounts of memory.