17 reservoir research
Fun recursive search of water until the rock bottom (or end of the
ground as we know it) is found. The input appears to be a long narrow
scan that contains random-sized buckets.

Drop water droplets one at a time and divide the droplets when a falling
one hits the ground; recurse those until the "droplet" has filled a
whole level again by hitting walls and letting the hit-the-ground
detection know about a wall in both sides (fill with water) or about a
cliff in at least one side (fill with water channel).