18b sustainable lumber collection
After some hundreds of magical minutes, the forest turns into a spiral
that spins around, generating trees on one side and munching them with
lumberyards that move towards them on the spiral tangent.

This spiral spin motion repeats; as a result, the total resource value
of this forest also repeats. Detect the period and extrapolate the
resource value at ridiculous amount of time from the start.

For my input, the repeating pattern has a length of 28 seconds; this
happens to hit the value after exactly 1000 iterations.