19b a new background process spinning
The program computes some kind of algorithm that takes a long time to
execute: after a careful decompilation into proper flow control, this
clearly sums up the numbers that divide r5, including itself.

The value in r5 is rather large, and looping large * large takes a long
time, but that is easily optimized. (Even faster might be to find the
prime factors and play with them, but this is faster to write.)