21 chronal conversion
An activation system for the time travel module appears to calculate a
some sort of sequence of random numbers. This time the tricky code is a
bit more involved (three nested loops) so I drew some control flow graph
diagrams for fun, too. Now I want to make a disasmer and cfg analyzer.