24 immune system simulator 20xx
The reindeer is protected by an immune system that fights a game of two
competing armies with rather straightforward target selection and
attacking rules. The input is interesting to parse though but easy
enough and fun with some regexing.

The attackers and targets are selected first to a vec of refs and then a
fight is carried out. Can't have mutable refs to the same vec that's
also read, so the unit count that gets modified when an army is wounded
is stored in a Cell. Not sure if this is the best way; alternatively we
could copy the target counts for the next iteration and then update on
one go after the whole fight. Vec indexes could also be used but those
feel dangerous like dangling pointers.

23b is still pending in another branch for now. I need to start working
on my off-by-one thinking this year.