24b more milk and cookies would help
We're giving the immune system a boost by increasing the unit attack
points by some constant number; do a linear search (which ends up at
exactly a boost of 42 for my input). The game isn't slow to calculate at
all so binary searching wouldn't be useful.

Now the game has to determine whether it would end up in a full tie such
that the two armies would not be able to yield enough damage to kill any
units after some point; this would hang the iteration if not detected.
In such cases the reindeer might not die from the infection but would
never heal either.

There's a catch: the rules don't say anything about not doing attacks if
the attack would not kill any units so those need to still consume some
targets or else the selection order would go wrong. Filtering like that
and checking whether there are any attacks would not work. Tracking the
number of fallen units is working though.