25 four-dimensional adventure
This adventure wasn't a significant one with the relatively small input
size. New constellations are formed by trying each point in the
initially ungrouped list of points in the spacetime. Each point needs to
be tested against at most all other points currently processed so this
is just O(n^2). Less than a million comparisons is nothing.

Each new point can be close to several already formed constellations so
for each new point all currently formed constellations are tested, and
joined together if it works (and removed from the list, because they
become new, different constellations). At the end of the testing round
this new constellation joins the list again.