23b figure out to actually teleport
When the nanobots activate, the place to be is the coordinate which is
in the range of the largest number of nanobots. Ties are split by
picking the one closest to origin.

This has probably dozens of ways to calculate, but I'm doing a BSP
search in the discrete-coordinate space. The best density (in just some
coordinate) is found first, and then another search is performed that
also takes into account the manhattan distance between the coordinate
and origin.

Each bot has the shape of an octahedron. Curiously they're also
distributed in space into a shape that resembles an octahedron and there
is clearly an area with a higher number of bots than elsewhere.
Considering the relatively huge ranges of the bots, there's significant
overlap all over the space.

The search is a nice usual recursive BSP that subdivides AABBs in two
and prunes the sides that end up having a worse total density than the
currently known best point density. The total density (which is an upper
bound for a pointwise density in that AABB) is computed from the number
of bots intersecting a single AABB which ended up being very bug prone.
Hyperplane separation theorem to work, with some special cases in this
particular configuration but with a very naive implementation anyway.