6 universal orbit map
The input is a DAG of celestial objects orbiting other objects. The
number of direct orbits would be the number of objects; the indirect
orbit count comes from the objects orbiting their parents' parents etc.
recursively which happens to be just the distance to the universal
Center of Mass. Search the parents recursively (sort of inverse DFS) to
compute the hop count of each object.