6b orbital transfers for the way to santa
Travel in the graph (tree, actually) to find the object on both YOUr and
SANta's road to COM. The distance to santa is the sum of those two
paths.

This is a bit simpler because there's not really need to memoize the
distances because we're searching to COM just from two places. If the
paths would be terribly long, the "sanpath.contains(x)" might become too
expensive. A hashset might work there.