6b use the count map for hops to santa
The hop count to santa can be calculated from the count map after a full
search, so reuse the search from part 1 and also find the common knot
point by stepping both paths at the same time. This might be more
efficient in a real world universe with an input orders of magnitude
larger than in the given input of just ~1.7k objects.