7 amplification circuit
Five programs in a chain now, this is getting interesting. Output from
one amplifier controller goes into the input of another, until the
output goes to the thrusters.

No changes to the computer instruction set, but now the programs may
take more than one input value so the input is an iterator. Each program
appears to read exactly two values followed by one write, at least with
this sort of input range.

All inputs tested are the permutations of the possible phase settings:
[0, 1, 2, 3, 4]. The phases are stored as an "array" in bitfields just
for fun. The permutations are generated with Heap's algorithm borrowed
from Wikipedia.