7b rewire it into a feedback loop
The amplifier oscillates now with a different input (with phase settings
5..9): the output of the last amp is connected to the first one's input.
The values grow quite a bit, so use i64.

Now that the programs run longer and they need to live "concurrently"
before stopping, the "computers" need to be executed only a bit at a
time each; this is implemented by running them until the next output
occurs (to be injected into the next one in the loop).

Both the initialization vector from the phase settings and the first
input from the previous amp (or 0 for the very first run) are read
before the first output, so the input still needs to be an iterator
(unless the execution would break for each input; then they could be
"primed" with just the phases).