9 sensor boost
With the Basic Operation Of System Test and all, the computer must now
understand "relative base" addressing when not in immediate mode: the
computer holds a single base register that is summed to positional
addressing. A new opcode increments or decrements the base. Supporting
multiple outputs is also useful at least for the quine test, although
the puzzle answer is just one keycode value.