10b vaporize with a giant laser
The only solution is awesome. Is it *complete* vaporization though if
the asteroid on which the monitoring station is set up is not vaporized?
The spec doesn't say anything about that.

The pulverization occurs clockwise with a spinning gun so that shadowed
objects are destroyed on the next round. In an infinitely large space
with asteroids very close to each other when thinking solid angles, we
might want to use fractional numbers (tan alpha = dy / dx). In this
silly exercise just atan2 will do.

On each round the zapped objects are tracked and the objects shadowed
them are hidden so that removal is left for the next round.