beautify double-collecting flat_map in 10b
Thanks cyndis for pointing out about into_bytes().

The problem seems to be that the inner closure that would return the
line iterator owns the original by-value param string, and the closure
would return an iterator that just would hold a reference to the string
with bytes(). So the closure needed to collect the filter-mapped
coordinates already.

With into_bytes() we get to own the string's internal byte vector and an
into_iter() takes ownership of that, so the returned iterator no longer
would hold a reference that would go out of scope when returning from
the closure.

The lines could also be collected first so that the outer loop would
iterate their references and the strings would stay alive. That I also
tried first but it's even uglier than the early coordinate collect.