11 space police
Turns out that even in space you need a license plate, or a "clearly
visible registration identifier". Looks like the intcode computer would
soon need to move to a "library" to avoid copypasta; this time it's
unchanged from the previous one and it's used to guide a robot. It's
nice that every puzzle solution is contained though.

The robot walks on a grid of panels and paints them. Let's see how many
panels it has eventually marked by tracking the painted ones in a
hashmap (they're painted black or white). The input to the program
varies per iteration depending on what it's painted before.