12b history repeats itself
The history of the moons appears to repeat itself, just like the AoC.
The simulation will repeat from the beginning at an absurdly late point
in time. Like before, the trick is to notice that each position and
velocity in each axis repeats itself much earlier than the total
sequence. The whole sequence then repeats itself at the least common
multiple of the periods of all repeating series.

A period of a single value is brute-force searched by running the
simulation a reasonably long time and looking for overlapping sequences.
Very small sequences would be false positives, but such anomalies don't
seem to occur.