13 care package
The ship must be very large if it's easier to build an arcade cabinet
than to travel to the other end to play it. Intcode copypasta again,
with a Breakout game loop probably to come on top. For now just render
the screen once and count the blocks to break.