13b insert coin, wiggle joystick
Play for free and cheat with an AI to punch the joystick into the ball's
general direction. If there's some kind of neat scoring algorithm based
on the bricks' positions, or perhaps a PRNG, we don't use it directly
but just play the game until it stops with a final score. The score
displayed in the magical out-of-screen segment display is detected and
printed with the game screen at each iteration for sweet animation.