improve 14b with binary search ish thing
Start trying the fuel production with a big number and split it in half
each time the source matter would not last long enough. This turns the
silly over-a-minute computation into a few milliseconds, although
there's probably an analytical solution as well.

The chemistry loop manipulates the matter stores early, so when the
result says that there's not enough ore when asking for too much, the
stores break and need to be thrown away. The digital landfill at
/dev/null receives these bits happily for recycling.