15 oxygen system
A remote controlled droid is guided by the intcode computer (btw.
should reverse-engineer these programs, they must be interesting!). The
droid reports status codes that encode how the failed section of the
flat spaceship looks like.

Control the droid to travel through the whole isolated area for a
complete map in dfs fashion. With the complete map, run a bfs for the
shortest distance to the ruptured oxygen tanks in the supply unit.