15b oxygen gradually fills the area
The air supply speed is one minute to adjacent locations which in an
open or branching space would be kind of exponential? Maybe the density
is not constant. Anyway this is again solved with bfs: start from the
supply coordinate and fill the whole map. How many minutes it will take
to fill the area with oxygen is then the maximum distance in the search.