16 flawed frequency transmission
FFT in some books stands for fast fourier transform and this given
pattern looks vaguely like the butterfly diagram but it's probably not
even close. For now the input list of digits is small enough that naive
iterator fun is good rust practice and fast enough. New digits are
computed one by one from the dot product with the pattern that basically
just forgets and moves around various digits.