16b the bigger real signal is easier to compute
The real signal is the original concatenated 10k times and the position
of the eight-digit message is not at the beginning but at an offset
described by the first seven digits. The offset makes the computation
relatively easier.

The offset is quite large, so the pattern becomes a string of zeros
followed by a string of ones that can be used in O(n) time instead of
O(n^2) time, speeding up quite a bit when the repeated signal length is
6,5 million digits. Each digit computation refers to either no digits or
all the digits to the right from it, using their sum, so fold the sum
from the right and store the value of each resulting digit.