add intcode disasm tool
Initial commit for a minimal disassembler for intcode. Produces a
convincing cfg picture on basic blocks and a nice textual dump for day
15's input. Likely needs a bunch of work to support relative addressing,
self-modifications and inline data.

Day 15 is the oxygen repair task with the remote-controlled droid.
That's approximately when this thing was developed. The intcode for the
droid reports walls about the maze and the code is of reasonable size so
that looks like an okay exercise to start with.

(Delete the asm "docs" in day 15; the branch and test instructions were
accidentally swapped.)