17b notify all other robots on the scaffold
Override the randomness of the vacuum robot and substitute new rules.
The robot needs to access every part at least once, and the map seems to
have a few repeating structures and it's not too big, so running a
compression algorithm manually in a text editor is enough to solve this.
An automated version is coming soon.

The algorithm packs the repeats into three "functions" called by the
"main routine". Given the rules, the robot starts to animate its
movement and it finally prints out the amount of dust collected.