19 tractor beam
The tractor pulling competition at Triton needs to get a good picture of
the beam. The beam is inspected with a drone system that is accessed
with an intcode program. The program inputs coordinates and outputs
beaminess. Scan the required 50x50 area and compute how much beam it
contains. A debug dump is provided as well.

The beam looks funny; it has a gap between the origin and the first
visible beam pixel. After that the beam looks contiguous. Perhaps the
beam is just so hot that it's invisible, like a gas torch.