18 many-worlds interpretation
The beam forces us to land next to an underground vault consisting of
tiles of open space, walls, keys, or doors. An optimal path (traveling
salesman but harder because keys??) is required to be found because all
the keys are needed to disable the tractor beam.

With the grid input, an adjacency list is first collected with a raw
BFS, keeping track of which keys have been collected and which doors are
passed through. With these edges, another (dijkstra) search is executed
starting from the player position.

With a global optimal distance list that encodes which keys are held at
each time, the resulting search will find all destinations where all
keys are held. The shortest of these is the answer.

Even after much cleanup after the solution day and some optimization for
bitmap key stores after finding the initial solution in minutes, this is
very slow (at 18.5 seconds for the puzzle input, or 1.3 seconds for the
big sample) so something must be wrong.