22 slam shuffle
Given a deck of space cards and a list of shuffling instructions, follow
the instructions for our deck of 10007 cards (prime number!) and find
the resulting position of the card 2019 that originally was at place
2019 when the deck came from the factory.

The second part of the day must be a huge repetition of the given
shuffle process with the same deck or something. After running the whole
sequence, each card ends up into another position that's consistent on
each iteration so the techniques could be skipped. With this small
numbers, the first part is easily done with a whole vec of cards; that
would serve as a gold standard to compare against with more complex
puzzle.