22b the 3d printers combine their power
A giant factory order deck appears and it's worthy of shuffling. It's
shuffled a giant number of times. After shuffling, an inverse lookup is
needed to find the original card that sits at a specific position.

The shuffling techniques can be represented as a*cardpos+b with constant
a and b for each tech, in modular arithmetic. They can also be combined
into just one shuffle iteratively. This single shuffle can then be
exponentiated mod cardsize, and when the reverse shuffle per destination
is expod, the original card that gets into position 2020 is found.

Some numbers appear to increase quite a bit, so use i128 for everything.
Even that isn't enough so the result is found by cheating a bit (FIXME),
outputting the final multiplication source values so it can be pasted to
a working calculator. Exponentiation by squaring for the whole (ax+b)^n
might help there as well; intermediate values in mod deck size would
stay smaller.