23 category six
The cat6 cable disaster broke communication on the ship, so we'll
rebuild it. Day 7 was quite fun already with the amplifier cascade and
should be revisited for concurrency overengineering, but today the
number of computing machines is 50 in terms of network interface
controllers. Build a network of the controllers and see what kind of
packet gets sent to address 255 (out of 50).

This might translate naturally to simulating the NICs in multiple
threads; for now there's a scheduling loop to traverse them one by one,
send if there's any packets pending to a NIC's queue, and receive if the
NIC wants to say something. The whole simulation stops when the address
255 would receive something.

The earlier intcode mainloops would not work here as is due to the
relative dependencies between the NICs except if the loop would schedule
out after a number of idle reads. Turning the sim mainloop into a Rust
generator might be also interesting. For now the programs are
interleaved step by step.