23b not always transmitting for power management
The "NAT" device which is entirely unlike network address translation
sends out packets when it detects that the network has been idle. Rework
the NAT tracking for a whole packet instead of i64 and track the count
of -1s read by the NICs. Figure out which value y is sent twice by the
NAT by returning the current packet held by the NAT (for upcoming power
mgmt transmissions and the first puzzle) and the last delivered packet
(for the second puzzle).

The NICs, their queues, idle flags, and perhaps the NAT state could be
thought as single state that maybe should be stored in one mutable
struct instead of passing them around separately. Whatever.