disregard door nodes in day 18 search
Of course the search space needs to be done only for key nodes in the
graph; the doors are only useful for looking whether keys behind them
can be accessed, but the order in which the doors taken is not
important. Skipping them reduces the search dimension from 2*26 to 26
(size of the alphabet) resulting in a ~10x speedup. That was so stupid.