24 planet of discord
Eris is infested with bugs that multiply and die in a cellular
automaton. As the whole area is just five by five cells, it fits in 32
bits so that's how the state is represented. Interestingly enough, this
bitmap state is also the layout's "biodiversity rating" that is the
puzzle answer for the first layout that appears twice. Run the
simulation, inserting each state to a set, looking for the duplicate.

Perhaps the second part will be a cycle puzzle again because of the
repetition hint?

As the state is a bitmap, the life operations could be encoded as bit
ops instead of looking at the cells naively one by one. Might revisit
that later.