24b eris is an old plutonian settlement
With increased knowledge about the recursively-folded space, the
cellular life computation is slightly more involved at the borders of
each recursion level. The 5x5 grids cascade like the cross sections of a
funnel. Now that the puzzle is about the number of bugs and that the
scan lasts only 200 minutes, 100 plus one recursion levels are enough to
encode the whole infinite recursion; when bugs initially exist only at a
middle level, this is the amount of cells that can possibly be infested
at the end.

The code repetition is nasty but manageable. Now the simulation counts
the sum of cells in each neighbour because the next level in the center
is five neighs that may be occupied.