bugfix and optimize day 18 shortcut edges
A problem with the shortcut edges that pick up keys along the way
through a corridor is that those picked up key nodes are not saved in
the distance graph; this causes them to be traversed too. Apparently the
number of shortcut edges is significant (of course) so not doing them
results in a 10x speedup. Now the solution speed is reasonable.

It would be nice to use the shortcuts though. This will need double
checking; maybe the extra nodes can be marked somehow or just skipped
from the search after being collected on another way.