sort day 18 heap by collected keys before labels
After comparing the distances for the priority queue, the contents of
the node data is compared (automatically) to make ordering consistent.
It appears to make sense to compare the number of keys collected
(keychain bitmap popcount) before the node label (which doesn't have any
particular meaning): when shorter paths are preferred, 3575300 nodes are
visited. When longer are preferred (by comparing the keys the other
way), 4626304 (30% more) are visited.

Intuitively it would make sense to prefer longer keychains (longer
paths) when the distance is otherwise equal because longer seems like
closer to completed, but on the other hand, there' no guarantee that
those longer paths would be ultimately better even if they greedily seem
like so. Looking at smaller keychains skips bad paths from the
beginning. Compare to DFS vs. BFS.

This cuts the search execution time almost in half.