prefer keys over distances in day 18 heap
Step a bit closer to something like A* search by using a strange
heuristic that appears to work and produce correct results with just
2.7% of the node visits of the previous commit. This is ridiculous. Make
the priority queue prefer small keychains over short walking distances.

The nodes in the search space are pairs of which key we're standing on
and which keys we're holding. With the funny distance map that's indexed
by the physical key only which isn't the whole node, there has to be the
extra check to only visit keys that we aren't holding yet.

Perhaps it is this combined with the keychain length ordering that
reduces the search space so much? Or perhaps this doesn't make sense and
is just a hacky fix for something that went way wrong from the start?