25 cryostasis
Faint life signature is found on the ship and the droid that gets sent
there communicates in ASCII Intcode. This ends up being very much like a
dungeon adventure game in text mode; the interesting bit is the
pressure-sensitive floor inside this Reindeer-class starship that we
need to trick into thinking that the droid's identity is that of some
deer or a small crew member.

Collect a map by DFSing the droid and print the map in graphviz format.
While traversing the ship for the map, pick up also all items except
those funny ones that would end the game.

When the DFS ends at the hull breach point again, use the map to
navigate to the security checkpoint that precedes the pressure-sensitive
floor. Try all 256 possible combinations of the eight collected items
until the right one is found; save the game by copying the computer
state so that the traversal from the beginning can be skipped.

With the right amount of items, step into the sensing floor for the last
time to pass on and find Santa who radios us the airlock password.

This code is a big mess and needs a cleanup to work all the way at once.