18b not one vault but four
Upon arrival, the vault looks like four different ones that have magical
keys and doors that are linked together even if the key is held by
someone else. To get all the keys, now four robots are used
simultaneously. The mission is again to figure out the shortest route to
pick up all keys; this means a somewhat different internal state, but
otherwise the "dijkstra" search is similar.

Just to get the puzzle star, copypaste the search function to be cleaned
up later. Adjust the previously scanned map for the middle walls that
separate the quadrants, deploy four searchy bots, and begin the search
with a state that includes all four positions.

The search edges are compiled such that one bot takes steps at a time
into neighbors that it has. If a bot is locked up waiting for keys to be
collected by others, it has no edges.