20b ancient plutonians built recursive spaces
The connections are not actually portals but doors into inner and outer
dimensions of the same structure (day 24b looks much like this). The
"portal" pairs are on the inner and outer sides of the donut structure.
Add a hacky flag to the main search to support recursion mode such that
the portals change a depth number in the state. The original teleporting
search goes on level 0 only.

Fake the search a bit and don't step over 100 deep. Maybe the number of
portals and the number of steps between them could work as a heuristic.