disasm: add formatting for opcode types
Now that real types exist, implement the string display for them instead
of hardcoding the description strings when reading the machine language.

Alignment or any other interesting formatting is not implemented
properly yet. When aligning the whole instruction descriptions in the
assembly listing, the descriptions are first formatted and the string
from that format is padded. This is suggested in the manual of std::fmt.

The lowest level of integers in SourceParam and DestParam are aligned
correctly.