finish day 15 asm decompiling
Update the variable names for day 15 and complete the decompiled C-like
source code listing. Add a script for extracting the map from the
program.

Identified variables:

- current x and y position
- whether those positions are odd or even
- y divided by two
- next values for the above (starting with "n")
- input and output
- temp storage

The oxygen supply lies at (33, 35) in my input. All world at x=0, y=0,
x=40 or y=40 are walls. The map lookup uses self-modifying code for
loading the correct map tile from memory for comparison.

Now the map logic is crystal clear. The map is 39 chars wide and 39
chars tall, but there's only 20 rows worth of map data. Each data row
encodes two rows' worth of spaceship with a clever trick.

When both x and y are odd, the space is free to walk on. This means that
every second row is 50% free, other data comes elsewhere:

.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?. (y=1)
??????????????????????????????????????? (y=2)
etc.

When both x and y are even, the space is walled. This also means that on
the rows that don't map to the previous rule, 50% is wall:

??????????????????????????????????????? (y=1)
?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#? (y=2)
etc.

These two non-overlapping rules result in the below grid that is
overlaid in front of the actual "bitmap", with map data read from "?":

123456789... (x)
.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?. (y=1)
?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#? (y=2)
.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?.?. (y=3)
?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#?#? (y=4)
etc.

When only one of the coordinates is odd, the wall data is read from an
array that's obfuscated with random data; a threshold value of 37
separates free space and walls. The stored half of the y coordinate is
used for the table lookup because each row in the map data represents
interleaved data for two rows.