2 password philosophy
We can't fix the password database yet with this, but the corporate
policy can be used to validate the mess. Easy enough to look at the
character counts for now. The pair param stuff with filter is not
beautiful but neither is a wrapper lambda. This again looks like an
exercise in parsing an input of something more than just numbers because
the puzzle itself is so trivial.

I kind of want to keep these solutions small and independent but regex
is again where I draw the line. Splitting strings is worse than
splitting hairs. Use this to build:

  rustc -L /path/to/regex/target/release/deps 2.rs

Or maybe some proper parsing library could be interesting to look at,
too.